ERC 1155 renewal

+++ BlockchainGameProgramming/erc-1155/contracts/Marketplace.sol
@@ -7,7 +7,7 @@ contract Marketplace {
 
     IERC1155 private _token;
 
-    mapping (uint => uint) price;
+    mapping (uint256 => uint256) price;
 
     constructor (IERC1155 token) public {
         require(address(token) != address(0));
